requirements.txt
django==3.2.16
django-environ==0.9.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
Pillow==9.3.0
django-storages==1.13.1
django-ckeditor==6.3.2

requests==2.28.1
djoser==2.0.4
social-auth-app-django==4.0.0
social-auth-core==4.3.0
PyJWT==2.5.0
argon2-cffi==21.3.0
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
whitenoise==5.2.0
boto3==1.24.92
django-storages==1.13.1
django-admin
dj-database-url==1.0.0
python-slugify==6.1.2
django-jazzmin==2.6.0

